+++ src/fisca/calculettes/calculette.js
@@ -1,0 +1,6 @@
+var BaseCalculette = (function () {
+    function BaseCalculette(params) {
+    }
+    return BaseCalculette;
+})();
+exports.BaseCalculette = BaseCalculette;