+++ src/fisca/calculettes/calculettes.ts
@@ -1,3 +1,5 @@
+import { ImpotRevenuCalculette } from './impot-revenu/impot-revenu';
+
 export interface ICalculette {
     calculer: any; // la fonction qui appel les calculs
     context: any;
@@ -10,7 +12,7 @@ export interface IParamsCalculette {
 export abstract class BaseCalculette {
 
     constructor(params: any){
-        
+
     }
 
 }