+++ src/fisca/fisca.js
@@ -1,0 +1,2 @@
+var Calculettes = require('./calculettes/calculettes');
+exports.Calculettes = Calculettes;