+++ src/fisca/calculettes/calculettes.js
@@ -1,6 +1,4 @@
-var BaseCalculette = (function () {
-    function BaseCalculette(params) {
-    }
-    return BaseCalculette;
-})();
-exports.BaseCalculette = BaseCalculette;
+var impot_revenu_1 = require('./impot-revenu/impot-revenu');
+exports.ImpotRevenuCalculette = impot_revenu_1.ImpotRevenuCalculette;
+var plus_value_immobiliere_1 = require('./plus-value-immobiliere/plus-value-immobiliere');
+exports.PlusValueImmobiliereCalculette = plus_value_immobiliere_1.PlusValueImmobiliereCalculette;